requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos
boto3
pandas
apache-airflow-providers-snowflake
apache-airflow-providers-amazon

